Internship-Digit-main/App/requirements.txt
numpy==1.24.3
Pillow
streamlit==1.27.0
tensorflow==2.13.0
tensorflow_intel